fix: save major account balance value in database when getting from bridge

+++ src/main/java/app/bpartners/api/endpoint/rest/mapper/TransactionRestMapper.java
@@ -14,7 +14,7 @@ public class TransactionRestMapper {
   public Transaction toRest(app.bpartners.api.model.Transaction internal) {
     Transaction transaction = new Transaction()
         .id(internal.getId())
-        .amount(internal.getAmount().getCents().intValue())
+        .amount(internal.getAmount().getCents())
         .label(internal.getLabel())
         .paymentDatetime(internal.getPaymentDatetime())
         .reference(internal.getReference())
